remove `Actor` and `ActorSystem`, keep only `unstable/machine` with runtime included
