+++ packages/effect/package.json
@@ -30,7 +30,6 @@
     "./package.json": "./package.json",
     ".": "./src/index.ts",
     "./testing": "./src/testing/index.ts",
-    "./unstable/actors": "./src/unstable/actors/index.ts",
     "./unstable/ai": "./src/unstable/ai/index.ts",
     "./unstable/cli": "./src/unstable/cli/index.ts",
     "./unstable/cluster": "./src/unstable/cluster/index.ts",
@@ -39,6 +38,7 @@
     "./unstable/eventlog": "./src/unstable/eventlog/index.ts",
     "./unstable/http": "./src/unstable/http/index.ts",
     "./unstable/httpapi": "./src/unstable/httpapi/index.ts",
+    "./unstable/machine": "./src/unstable/machine/index.ts",
     "./unstable/observability": "./src/unstable/observability/index.ts",
     "./unstable/persistence": "./src/unstable/persistence/index.ts",
     "./unstable/process": "./src/unstable/process/index.ts",
@@ -69,7 +69,6 @@
       "./package.json": "./package.json",
       ".": "./dist/index.js",
       "./testing": "./dist/testing/index.js",
-      "./unstable/actors": "./dist/unstable/actors/index.js",
       "./unstable/ai": "./dist/unstable/ai/index.js",
       "./unstable/cli": "./dist/unstable/cli/index.js",
       "./unstable/cluster": "./dist/unstable/cluster/index.js",
@@ -78,6 +77,7 @@
       "./unstable/eventlog": "./dist/unstable/eventlog/index.js",
       "./unstable/http": "./dist/unstable/http/index.js",
       "./unstable/httpapi": "./dist/unstable/httpapi/index.js",
+      "./unstable/machine": "./dist/unstable/machine/index.js",
       "./unstable/observability": "./dist/unstable/observability/index.js",
       "./unstable/persistence": "./dist/unstable/persistence/index.js",
       "./unstable/process": "./dist/unstable/process/index.js",
