+++ packages/effect/src/unstable/machine/index.ts
@@ -7,14 +7,4 @@
 /**
  * @since 4.0.0
  */
-export * as Actor from "./Actor.ts"
-
-/**
- * @since 4.0.0
- */
-export * as ActorSystem from "./ActorSystem.ts"
-
-/**
- * @since 4.0.0
- */
 export * as StateMachine from "./StateMachine.ts"
