+++ packages/effect/src/unstable/machine/internal/stateMachineModel.ts
